pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sts2-rl"
version = "0.1.0"
description = "Reinforcement learning toolkit for Slay the Spire 2 built on a local sts2-cli fork"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gymnasium==0.28.1",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "flask>=3.0.0",
    "pygame>=2.5.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]
rl = [
    "pyyaml>=6.0",
    "stable-baselines3>=2.2.0",
    "shimmy>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
